+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## ChangeLog
 
-### 2.x.x
+### 2.0.8
 * Added a function that returns all the elements that corresponding with a tag name
 * Delete the `addstyle` option
 * Added the `deleteExceedCharacter` option